requirements.txt
## `requirements.txt`

openai
requests
pyyaml
yagmail
